Make the LFO correctly behave as a full sine wave
The LFO was previously only going through the positive values of a sine wave.
That creates a problem with vibrato, where the pitch oscillates between
the starting value and a higher one, effectively raising the pitch of
the note.
A correct vibrato oscillates both above and below the starting pitch,
thus maintaining the perceived pitch centered on the same original note.